redis-storage: split test files
no functional change
